+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
-
 ## 2.5.6
 
 ## Enhancements
+
+- Add support for rerouting back button if enabled in paywall settings
+  - Handled by `SuperwallOptions.PaywallOptions.onBackPressed`, which enables you to consume the back press or let the SDK consume it
 - Add support for redeeming web entitlements with Paddle
 
 ## Fixes